requirements.txt
requests
braintree
python-telegram-bot==13.15
